+++ src/Storage/Storage/File/Cmdlet/NewAzureStorageFileSasToken.cs
@@ -95,6 +95,10 @@ namespace Microsoft.WindowsAzure.Commands.Storage.File.Cmdlet
         [ValidateNotNullOrEmpty]
         public string DelegatedUserObjectId { get; set; }
 
+        [Parameter(Mandatory = false, HelpMessage = "Optional. The delegated user tenant id in Azure AD. This parameter can only be specified when input Storage Context is OAuth based.")]
+        [ValidateNotNullOrEmpty]
+        public string DelegatedUserTenantId { get; set; }
+
         [Parameter(
             Mandatory = false,
             HelpMessage = "Permissions for a file. Permissions can be any subset of \"rwd\".",
@@ -181,6 +185,10 @@ namespace Microsoft.WindowsAzure.Commands.Storage.File.Cmdlet
                 {
                     throw new ArgumentException("DelegatedUserObjectId can only be specified when input Storage Context is OAuth based without using SAS token.", "DelegatedUserObjectId");
                 }
+                if (this.DelegatedUserTenantId != null)
+                {
+                    throw new ArgumentException("DelegatedUserTenantId can only be specified when input Storage Context is OAuth based without using SAS token.", "DelegatedUserTenantId");
+                }
             }
 
             // Get share saved policy if any
@@ -191,10 +199,10 @@ namespace Microsoft.WindowsAzure.Commands.Storage.File.Cmdlet
             }
 
             //Create SAS builder
-            ShareSasBuilder sasBuilder = SasTokenHelper.SetShareSasBuilder_FromFile(fileClient, identifier, this.Permission, this.StartTime, this.ExpiryTime, this.IPAddressOrRange, this.Protocol, this.DelegatedUserObjectId);
+            ShareSasBuilder sasBuilder = SasTokenHelper.SetShareSasBuilder_FromFile(fileClient, identifier, this.Permission, this.StartTime, this.ExpiryTime, this.IPAddressOrRange, this.Protocol, this.DelegatedUserObjectId, this.DelegatedUserTenantId);
 
             //Create SAS and output it
-            string sasToken = SasTokenHelper.GetFileSharedAccessSignature((AzureStorageContext)this.Context, sasBuilder, generateUserDelegationSas, CmdletCancellationToken);
+            string sasToken = SasTokenHelper.GetFileSharedAccessSignature((AzureStorageContext)this.Context, sasBuilder, generateUserDelegationSas, CmdletCancellationToken, this.DelegatedUserTenantId);
 
             // remove prefix "?" of SAS if any
             sasToken = Util.GetSASStringWithoutQuestionMark(sasToken);